post

+++ resources/views/microposts/microposts.blade.php
@@ -20,12 +20,15 @@
                             {!! Form::open(['route' => ['microposts.destroy', $micropost->id], 'method' => 'delete']) !!}
                                 {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
                             {!! Form::close() !!}
-                            {!! Form::open(['route' => ['favorites.favorite', $micropost->id], 'method' => 'post']) !!}
-                                {!! Form::submit('favorites favorite', ['class' => 'btn btn-danger btn-sm']) !!}
-                            {!! Form::close() !!}
-                            {!! Form::open(['route' => ['favorites.unfavorite', $micropost->id], 'method' => 'delete']) !!}
-                                {!! Form::submit('favorites delete', ['class' => 'btn btn-danger btn-sm']) !!}
-                            {!! Form::close() !!}
+                            @if (Auth::user()->is_favorite($micropost->id))
+                                {!! Form::open(['route' => ['favorites.unfavorite', $micropost->id], 'method' => 'delete']) !!}
+                                    {!! Form::submit('Unfavorite', ['class' => 'btn btn-danger btn-success btn-sm']) !!}
+                                {!! Form::close() !!}
+                            @else
+                                {!! Form::open(['route' => ['favorites.favorite', $micropost->id]]) !!}
+                                    {!! Form::submit('Favorite', ['class' => 'btn btn-light btn-sm']) !!}
+                                {!! Form::close() !!}
+                            @endif
                         @endif
                     </div>
                 </div>
